requirements.txt
djangorestframework
pendulum==2.0.5
python-dateutil==2.8.1
pytzdata==2019.3
six==1.14.0